pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-cloud-ops-agent"
version = "0.1.0"
description = "Intelligent AIOps agent for multi-cloud resource management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Team", email = "team@example.com"}
]

dependencies = [
    "kubernetes>=28.0.0",
    "boto3>=1.34.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-identity>=1.15.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "gitpython>=3.1.40",
    "jinja2>=3.1.3",
    "apscheduler>=3.10.4",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

localstack = [
    "localstack>=3.0.0",
    "awscli-local>=0.21",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
